% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_MedianRentalPrice_2Bedroom}
\alias{Zip_MedianRentalPrice_2Bedroom}
\title{Zipcode Level Median Rental Price for 2-Bedroom}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57227 rows and 8 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_MedianRentalPrice_2Bedroom
}
\description{
Zipcode Level Median Rental Price for 2-Bedroom
}
\keyword{datasets}
