% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_MedianRentalPrice_AllHomes}
\alias{Zip_MedianRentalPrice_AllHomes}
\title{Zipcode Level Median Rental Price for All Homes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133144 rows and 8 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_MedianRentalPrice_AllHomes
}
\description{
Zipcode Level Median Rental Price for All Homes
}
\keyword{datasets}
