% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_PriceToRentRatio_AllHomes}
\alias{Zip_PriceToRentRatio_AllHomes}
\title{Zipcode Level Price-To-Rent Ratio Data for All Homes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1069200 rows and 9 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_PriceToRentRatio_AllHomes
}
\description{
Zipcode Level Price-To-Rent Ratio Data for All Homes
}
\keyword{datasets}
