% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_Zhvi_AllHomes}
\alias{Zip_Zhvi_AllHomes}
\title{Zipcode Level Zillow Home Value Index (ZHVI) for All Homes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3967266 rows and 9 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_Zhvi_AllHomes
}
\description{
Zipcode Level Zillow Home Value Index (ZHVI) for All Homes
}
\keyword{datasets}
