% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_Zri_AllHomes}
\alias{Zip_Zri_AllHomes}
\title{Zipcode Level Zillow Rental Index (ZRI) for All Homes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1272880 rows and 9 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_Zri_AllHomes
}
\description{
Zipcode Level Zillow Rental Index (ZRI) for All Homes
}
\keyword{datasets}
