% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zip_Zri_MultiFamilyResidenceRental}
\alias{Zip_Zri_MultiFamilyResidenceRental}
\title{Zip_Zri_MultiFamilyResidenceRental}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 683600 rows and 9 columns.}
\source{
\url{https://www.zillow.com/research/data/}
}
\usage{
Zip_Zri_MultiFamilyResidenceRental
}
\description{
Zip_Zri_MultiFamilyResidenceRental
}
\keyword{datasets}
